create ui_utils to remove utils dependency on html
